requirements.txt
# Core dependencies for Dong Thap Land Change Analysis

# Google Earth Engine
earthengine-api>=0.1.374

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Geospatial data processing
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
gdal>=3.11.1

# Data processing
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0
contextily>=1.3.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Parallel processing
joblib>=1.3.0
dask>=2023.5.0

# Quality and formatting
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
